Move get_feature_job_logs to TileJobLogService and fix tests
